Split MPoly.jl.
