+++ src/generic/InvariantFactorDecomposition.jl
@@ -141,6 +141,14 @@ end
 #
 ###############################################################################
 
+@doc Markdown.doc"""
+    snf(m::FPModule{T}) where T <: RingElement
+
+Return a pair `M, f` consisting of the invariant factor decomposition $M$ of
+the module `m` and a module homomorphism (isomorphisms) $f : M \to m$. The
+module `M` is itself a module which can be manipulated as any other module
+in the system.
+"""
 function snf(m::AbstractAlgebra.FPModule{T}) where T <: RingElement
    R = base_ring(m)
    old_rels = rels(m)
